refactor monitoring run detail router
