+++ src/app/routers/dpm_command_center_monitoring.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Path, Query
+from fastapi import APIRouter, Query
 
 from app.contracts.dpm_command_center import (
     DpmCommandCenterForwardRequest,
@@ -58,28 +58,5 @@ async def list_monitoring_runs(
 ) -> DpmCommandCenterGatewayResponse:
     return await dpm_command_center_service().list_monitoring_runs(
         filters={"status_filter": status_filter, "limit": limit, "cursor": cursor},
-        correlation_id=correlation_id_var.get(),
-    )
-
-
-@router.get(
-    "/monitoring/runs/{monitoring_run_id}",
-    response_model=DpmCommandCenterGatewayResponse,
-    summary="Get DPM monitoring run",
-    description=(
-        "What: returns one manage-owned mandate monitoring run. When: use this for audit "
-        "drill-down from command-center latest-run or exception evidence. How: Gateway returns "
-        "the manage payload in a product envelope without changing health or exception truth."
-    ),
-)
-async def get_monitoring_run(
-    monitoring_run_id: str = Path(
-        ...,
-        description="Manage-owned mandate monitoring-run identifier.",
-        examples=["dmr_20260503_083000"],
-    ),
-) -> DpmCommandCenterGatewayResponse:
-    return await dpm_command_center_service().get_monitoring_run(
-        monitoring_run_id=monitoring_run_id,
         correlation_id=correlation_id_var.get(),
     )
